#148 use interfaces instead implementations for autowired fields
