--- tickers/src/main/kotlin/fund/cyber/markets/ticker/processor/HopTickerProcessorInterface.kt
@@ -1,4 +1,0 @@
-package fund.cyber.markets.ticker.processor
-
-interface HopTickerProcessorInterface {
-}
