--- tickers/src/main/kotlin/fund/cyber/markets/ticker/processor/TickerProcessorInterface.kt
@@ -1,5 +1,0 @@
-package fund.cyber.markets.ticker.processor
-
-interface TickerProcessorInterface {
-
-}
